Refactored MatchProfile to ViewProfile Screen
