+++ navigation/HomeStack.js
@@ -7,7 +7,7 @@ import EditProfile from '../screens/EditProfile';
 import ResetPassword from '../screens/ResetPassword';
 import Questionnaire from '../screens/Questionnaire';
 import Interests from '../screens/Interests';
-import MatchProfile from '../screens/MatchProfile';
+import ViewProfile from '../screens/ViewProfile';
 
 import Colors from '../constants/Colors';
 
@@ -54,9 +54,9 @@ const HomeStack = () => {
 				}}
 			/>
 
-			<Stack.Screen name="MatchProfile" component={MatchProfile}
+			<Stack.Screen name="ViewProfile" component={ViewProfile}
 				options={{
-					headerTitle: "Match Profile",
+					headerTitle: "View Profile",
 					headerBackTitle: "Back"
 				}}
 			/>
